Move seed calc to unique seed module
