+++ rng_tools/src/generators/gen3/mod.rs
@@ -1,5 +1,6 @@
 mod egg;
 mod feebas_sid;
+pub mod seed3;
 mod stationary;
 mod tidsid;
