+++ rng_tools/src/generators/gen4/seed4.rs
