+++ rng_tools/src/generators/gen4/seed_time.rs
@@ -1,4 +1,4 @@
-use super::{calc_ab, calc_seed};
+use super::seed4::{calc_ab, calc_seed};
 use crate::rng::mt::MT;
 use crate::{RngDateTime, get_days_in_month};
 use chrono::Duration;
