+++ rng_tools/src/generators/gen6/mod.rs
@@ -1,12 +1,12 @@
 mod id;
 mod mirage_spot;
 mod poke_radar;
-mod seed;
+mod seed6;
 mod stationary;
 mod utils;
 
 pub use id::*;
 pub use mirage_spot::*;
 pub use poke_radar::*;
-pub use seed::*;
+pub use seed6::*;
 pub use stationary::*;
