+++ rng_tools/src/generators/gen6/seed6.rs
